Added multipart support
The Client class now takes an optional files keyword argument that, when
provided, will use the MultipartSignedRequestFactory to build the
request instead of the normal SignedRequestFactory. The
MultipartSignedRequestFactory will use the request as normal to build
the signature and pass it along, but also include the files in the
multipart request (after the signature has been added). This allows the
server to be able to recreate the signature without having to know about
the files included in the multipart request.