Bug Fix: Corrected issue where the Client ALWAYS tried to json.dumps when the data type was application/json
If the data provided to get_response was already a json string, this would lead to duplicate escaping and break things. The use case is: I receive a json payload from one service and want to post it to another service.  In the client adapter I don't care what the json content is, I just need to pass it along.  The user shouldn't be forced to json.loads  the data just to turn around and have this generic client json.dumps it again.